Refactor walk registration response handling.
Updated response construction in `RegisterWalk` to utilize `LazyCreationResponse` and `EntityFrameworkInsert` for better url generation. When using the Storage.Insert side effect, it does not update the entity id. Thus, the redirect url never gets the correct id. With the custom ef insert class, it does.

Changed `WalkResponse` and `DogResponse` so they can be returned by the public classes.