Add command-line configuration and partial Program class
Added command-line arguments to the configuration pipeline to improve flexibility in runtime settings. It also allows the integration tests to influence the environment such as the directory where the settings are found.

Introduced a partial `Program` class to allow the integration tests to reference the `Program` class in the setup.